Move to enums
