+++ azure-kinect/src/lib.rs
@@ -1,8 +1,7 @@
 pub use calibration::Calibration;
 pub use camera::Camera;
 pub use capture::Capture;
-pub use color_resolution::ColorResolution;
-pub use depth_mode::DepthMode;
+pub use enums::{ColorResolution, DepthMode};
 pub use device::Device;
 pub use error::Error;
 pub use factory::{DebugMessageHandler, Factory, FactoryRecord};
@@ -14,9 +13,8 @@ pub use transformation::Transformation;
 pub mod calibration;
 pub mod camera;
 pub mod capture;
-pub mod color_resolution;
-pub mod depth_mode;
 pub mod device;
+pub mod enums;
 pub mod error;
 pub mod factory;
 pub mod fps;
